# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Motor Clínico - API para clínica médica integrativa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /protocolos:
    get:
      operationId: listarProtocolos
      tags: [protocolo]
      summary: Listar protocolos terapêuticos
      parameters:
        - name: categoria
          in: query
          schema:
            type: string
        - name: ativo
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de protocolos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Protocolo"
    post:
      operationId: criarProtocolo
      tags: [protocolo]
      summary: Criar protocolo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarProtocoloBody"
      responses:
        "201":
          description: Protocolo criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Protocolo"

  # Filas Operacionais
...
  /usuarios:
    get:
      operationId: listarUsuarios
      tags: [usuarios]
      summary: Listar usuários do sistema
      parameters:
        - name: perfil
          in: query
          schema:
            type: string
            enum: [enfermeira, validador_enfermeiro, medico_tecnico, validador_mestre]
        - name: unidadeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
    post:
      operationId: criarUsuario
      tags: [usuarios]
      summary: Criar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarUsuarioBody"
      responses:
        "201":
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
  /usuarios/login:
    post:
      operationId: loginUsuario
      tags: [usuarios]
      summary: Login do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
components:
  schemas:
...
    Protocolo:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: string
        categoria:
          type: string
        itens:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: integer
              itemNome:
                type: string
              quantidade:
                type: string
              observacao:
                type: string
        ativo:
          type: boolean
        criadoPorId:
          type: integer
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - ativo
        - criadoEm
    CriarProtocoloBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        categoria:
          type: string
        itens:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: integer
              quantidade:
                type: string
              observacao:
                type: string
        ativo:
          type: boolean
        criadoPorId:
          type: integer
      required:
        - nome
        - criadoPorId
...
    Usuario:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        perfil:
          type: string
          enum: [enfermeira, validador_enfermeiro, medico_tecnico, validador_mestre]
        unidadeId:
          type: integer
        unidadeNome:
          type: string
        ativo:
          type: boolean
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - email
        - perfil
        - ativo
        - criadoEm
    CriarUsuarioBody:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        senha:
          type: string
        perfil:
          type: string
          enum: [enfermeira, validador_enfermeiro, medico_tecnico, validador_mestre]
        unidadeId:
          type: integer
      required:
        - nome
        - email
        - senha
        - perfil
    LoginBody:
      type: object
      properties:
        email:
          type: string
        senha:
          type: string
      required:
        - email
        - senha
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        usuario:
          $ref: "#/components/schemas/Usuario"
      required:
        - token
        - usuario
...